GameLoop

+++ Assets/Scripts/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    public abstract void TakeDamage(int damage);
+    public abstract void TakeDamage(float damage);
 }
